requirements.txt
Babel==2.5.1
Django==1.11.6
django-phonenumber-field==1.3.0
phonenumberslite==8.8.4
pytz==2017.2
